core/_archive/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
defusedxml==0.7.1
Django==5.2.1
django-allauth==65.8.1
django-ckeditor==6.7.3
django-filter==25.1
django-js-asset==3.1.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.0
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
djangorestframework==3.16.0
dnspython==2.7.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
laces==0.1.2
openpyxl==3.1.5
pillow==11.2.1
pillow_heif==0.21.0
python-decouple==3.8
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wagtail==7.0
Willow==1.10.0
